+++ app/classes/web/routes/api/api_handlers.py
@@ -57,12 +57,14 @@ from app.classes.web.routes.api.servers.server.files import (
     ApiServersServerFileDownload,
     ApiServersServerFilesOperationHandler,
 )
-from app.classes.web.routes.api.crafty.upload.index import ApiFilesUploadHandler
 from app.classes.web.routes.api.servers.server.tasks.task.children import (
     ApiServersServerTasksTaskChildrenHandler,
 )
 from app.classes.web.routes.api.servers.server.tasks.task.index import (
     ApiServersServerTasksTaskIndexHandler,
+)
+from app.classes.web.routes.api.servers.server.tasks.task.run import (
+    ApiServersServerTasksTaskRunHandler,
 )
 from app.classes.web.routes.api.servers.server.webhooks.index import (
     ApiServersServerWebhooksIndexHandler,
@@ -71,6 +73,7 @@ from app.classes.web.routes.api.servers.server.webhooks.webhook.index import (
     ApiServersServerWebhooksManagementIndexHandler,
 )
 from app.classes.web.routes.api.servers.server.users import ApiServersServerUsersHandler
+from app.classes.web.routes.api.servers.server.upload import ApiServerFilesUpload
 from app.classes.web.routes.api.users.index import ApiUsersIndexHandler
 from app.classes.web.routes.api.users.user.index import ApiUsersUserIndexHandler
 from app.classes.web.routes.api.users.user.permissions import (
@@ -105,6 +108,8 @@ from app.classes.web.routes.api.crafty.stats.stats import ApiCraftyHostStatsHand
 from app.classes.web.routes.api.crafty.clogs.index import ApiCraftyLogIndexHandler
 from app.classes.web.routes.api.crafty.clogs.support import ApiCraftySupportIndexHandler
 from app.classes.web.routes.api.crafty.imports.index import ApiImportFilesIndexHandler
+from app.classes.web.routes.api.crafty.imports.upload import APIServerImportUpload
+from app.classes.web.routes.api.crafty.config.upload import APICraftyCustomizeUpload
 from app.classes.web.routes.api.crafty.exe_cache import (
     ApiCraftyJarCacheIndexHandler,
     ApiCraftySteamCacheIndexHandler,
@@ -382,7 +387,7 @@ def api_handlers(handler_args):
         ),
         (
             r"/api/v2/crafty/admin/upload/?",
-            ApiFilesUploadHandler,
+            APICraftyCustomizeUpload,
             handler_args,
         ),
         (
@@ -392,12 +397,12 @@ def api_handlers(handler_args):
         ),
         (
             r"/api/v2/servers/import/upload/?",
-            ApiFilesUploadHandler,
+            APIServerImportUpload,
             handler_args,
         ),
         (
             r"/api/v2/servers/([a-z0-9-]+)/files/upload/?",
-            ApiFilesUploadHandler,
+            ApiServerFilesUpload,
             handler_args,
         ),
         (
@@ -428,6 +433,11 @@ def api_handlers(handler_args):
         (
             r"/api/v2/servers/([a-z0-9-]+)/tasks/([0-9]+)/children/?",
             ApiServersServerTasksTaskChildrenHandler,
+            handler_args,
+        ),
+        (
+            r"/api/v2/servers/([a-z0-9-]+)/tasks/([0-9]+)/run/?",
+            ApiServersServerTasksTaskRunHandler,
             handler_args,
         ),
         (